cleanup fortran compiler code use size_t instead of uint

+++ bld/f77/wfc/c/kwlookup.c
@@ -2,7 +2,7 @@
 *
 *                            Open Watcom Project
 *
-* Copyright (c) 2002-2020 The Open Watcom Contributors. All Rights Reserved.
+* Copyright (c) 2002-2021 The Open Watcom Contributors. All Rights Reserved.
 *    Portions Copyright (c) 1983-2002 Sybase, Inc. All Rights Reserved.
 *
 *  ========================================================================
@@ -44,14 +44,14 @@
 #include "kwlookup.h"
 
 
-int KwLookUp( const char **table, int high, const char *id, uint id_len, bool exact )
-//===================================================================================
+int KwLookUp( const char **table, int high, const char *id, size_t id_len, bool exact )
+//=====================================================================================
 {
     const char  *key;
     const char  *ident;
     int         mid;
     int         low;
-    uint        kw_len;
+    size_t      kw_len;
 
     // index 0 is reserved
     // each keyword table has blank first item
